Support API resources where API group doesn't match go package name & fix authentication API group
Previously, the generator took the package name as the API group,
but the API group and the package name in recently introduced
groups are no longer the same. For example, the TokenReview resource
is in the "pkg/apis/authentication" package, but the API group is
"authentication.k8s.io" instead of just "authentication".